refactor(frontend): layout -> shared + icons
