+++ packages/frontend/components/Shared/Icons/ProviderIcon.tsx
