+++ packages/frontend/components/Shared/Icons/index.ts
@@ -1,0 +1,4 @@
+export { default as DiscordIcon } from "./DiscordIcon"
+export { default as NetworkIcon } from "./NetworkIcon"
+export { default as ProviderIcon } from "./ProviderIcon"
+export { default as TokenIcon } from "./TokenIcon"
