+++ packages/frontend/helpers/tokenService.ts
@@ -1,0 +1,2 @@
+export const getTokenImage = (symbol: string) =>
+  `/assets/images/protocol-icons/tokens/${symbol}.svg`
