+++ packages/frontend/pages/index.tsx
@@ -2,8 +2,8 @@ import { NextPage } from "next"
 import Head from "next/head"
 import Image from "next/image"
 
-import Footer from "../components/Layout/Footer"
-import Header from "../components/Layout/Header"
+import Footer from "../components/Shared/Footer"
+import Header from "../components/Shared/Header"
 import styles from "../styles/Home.module.css"
 
 const Home: NextPage = () => {
